Look for python version on server
We need to examine the python version on the server to work out which
python version to use. But we can't do that in _setup_path() as that is
before fabric knows what server we're talking to.

So we have a function to work it out that caches the result.