fab: allow for maintenance apache config
When taking the site down for maintenance, check if the file
<webserver>/<environment>-maintenance.conf exists (so maybe
apache/production-maintenance.conf ).  If it does then link that
file instead of the normal apache config while the site upgrade is
taking place.

If you write your new site config to use a static HTML file as a
maintenance page then you get a nice holding page while the site is
down.  With the correct set up you can reuse the live CSS files.

See 3iesite for an example set up.