Change deploy directory structure
The new directory structure is timestamp directories in server project
home, with the timestamp being the time that directory was deployed.  A
soft link named current/ will point at the version that apache will serve.
For backwards compatibility with apache config, a dev/ link will also be
created, pointing at current/

The old was timestamp directories in <server project home>/previous/ with
the timestamp being the time the directory was archived.  The current
deploy was in <server project home>/dev/

This is motivated by virtualenvs being created with full paths hard
coded, and not being easy to move.  The new structure means the
underlying directory names do not change.

Finally update the cookiecutter apache config to point to current rather
than to dev