Allow for case where server hasn't pulled branch yet
It's possible that you've just pushed a branch to the repo, but the
branch doesn't exist on the server yet.  In that case we need to run
`git branch -r` on the local machine to see what branches have been
pushed.

This (should) fix issue #4 - but need to actually test it.