Skip some tests if mysql root password not available
Normally we use sudo to read the password from /root/mysql_root_password
but if that isn't there we need to put a prompt up to read the password.
This won't work during tests, so if we can't read the file, then skip
the test.

Also move to unittest2, as the skip stuff is only in unittest in python
>= 2.7, but unittest2  backports it.