Add --noinput option to tasks.py
To not ask for the MySQL root password, and then use it in the tests.

This should mean that when fabric runs, it won't have added the noinput
option, so when tasks.py runs on the server it can ask for the mysql
root password, and that will be passed to the user by fabric.