Refactor the copy_setting() stuff in _setup_paths()
Extend the copy_setting() function so that it will throw an exception
if a required setting is not present, and if a default is not given it
will not set the env variable.

Then put the required settings and the optional settings in lists so
that it's easier to parse.

Finally delete a few settings that shouldn't be in project_settings.py
- verbose, revision