Mount ramdisk on a subdirectory of /dev/shm instead.
If /dev/shm is a symlink to /run/shm (as on Ubuntu 12.04), the mountpoint
check fails and the script mounts another ramdisk over /dev/shm, hiding the
original contents, which causes several desktop applications to misbehave.

This patch may not fix the mountpoint detection, but it should at least stop
it from messing up the rest of the system.