If a git merge is required then check if user wants to continue
as that means there are commits on the server that aren't on the branch
we're deploying.  The user may want to check and do manual merging if
required.

To test we use `git merge --ff-only` which won't do anything if it can't
do a pure fast-forward.  And if the user does want to continue then we
use the `--no-edit` option to ensure that the deploy doesn't get stuck
with fabric talking to vim ...